+++ src/Commands.cpp
@@ -266,6 +266,7 @@ static SeqStrings gCommandNames =
     "CmdSetInverseSearch\0"
     "CmdToggleToolbarPosition\0"
     "CmdToggleDjvuEngine\0"
+    "CmdCommandPaletteFavorites\0"
     "CmdNone\0"
     "\0";
 
@@ -523,6 +524,7 @@ static i32 gCommandIds[] = {
     CmdSetInverseSearch,
     CmdToggleToolbarPosition,
     CmdToggleDjvuEngine,
+    CmdCommandPaletteFavorites,
     CmdNone,
 };
 
@@ -780,6 +782,7 @@ SeqStrings gCommandDescriptions =
     "Set Inverse Search Command Line\0"
     "Toggle Toolbar Position\0"
     "Toggle DjVu Engine\0"
+    "Command Palette: Favorites\0"
     "Do nothing\0"
     "\0";
 // clang-format on