+++ src/Commands.h
@@ -262,7 +262,8 @@ enum {
     CmdSetInverseSearch = 451,
     CmdToggleToolbarPosition = 452,
     CmdToggleDjvuEngine = 453,
-    CmdNone = 454,
+    CmdCommandPaletteFavorites = 454,
+    CmdNone = 455,
 
     /* range for file history */
     CmdFileHistoryFirst,